Refactoring attempt 1
